guardian: organize S3 log modules
