+++ crates/hashi-types/src/guardian/log/messages/init.rs
@@ -1,0 +1,81 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use super::super::ObjectKeyPattern;
+use super::super::S3_DIR_INIT;
+use crate::bitcoin::BitcoinPubkey;
+use crate::guardian::GuardianInfo;
+use crate::guardian::GuardianPubKey;
+use crate::guardian::LimiterState;
+use crate::guardian::NitroAttestation;
+use crate::guardian::ShareID;
+use serde::Deserialize;
+use serde::Serialize;
+
+/// OI: operator_init
+/// PI: provisioner_init
+/// Init messages are expected to be logged in the following order:
+/// OIAttestationUnsigned -> OIGuardianInfo -> PIEnclaveFullyInitialized -> OAActivated.
+#[derive(Debug, Serialize, Deserialize)]
+pub enum InitLogMessage {
+    /// Attestation and signing public key posted in /operator_init
+    OIAttestationUnsigned {
+        attestation: NitroAttestation,
+        #[serde(with = "crate::guardian::serde_utils::guardian_pubkey")]
+        signing_public_key: GuardianPubKey,
+    },
+    /// Signed GuardianInfo logged in /operator_init (secret-sharing instance,
+    /// config_hash, encryption/BTC pubkeys). Boxed: much larger than the other
+    /// variants (`clippy::large_enum_variant`).
+    OIGuardianInfo(Box<GuardianInfo>),
+    /// Threshold reached — enclave BTC key reconstructed (happens once).
+    PIEnclaveFullyInitialized {
+        sharing_seq: u64,
+        share_ids: Vec<ShareID>,
+        enclave_btc_pubkey: BitcoinPubkey,
+    },
+    /// Operator activation succeeded and installed live serving state.
+    OAActivated {
+        #[serde(with = "hex::serde")]
+        state_hash: [u8; 32],
+        #[serde(with = "hex::serde")]
+        config_hash: [u8; 32],
+        sharing_seq: u64,
+        committee_epoch: u64,
+        limiter_state: LimiterState,
+    },
+}
+
+impl InitLogMessage {
+    pub const OI_ATTEST_UNSIGNED: &'static str = "01-oi-attestation-unsigned";
+    pub const OI_GUARDIAN_INFO: &'static str = "02-oi-guardian-info";
+    pub const PI_FULLY_INITIALIZED: &'static str = "03-pi-enclave-fully-initialized";
+    pub const OA_ACTIVATED: &'static str = "04-oa-activated";
+
+    pub fn object_key(&self, session_id: &str) -> String {
+        let suffix = match self {
+            InitLogMessage::OIAttestationUnsigned { .. } => Self::OI_ATTEST_UNSIGNED,
+            InitLogMessage::OIGuardianInfo(_) => Self::OI_GUARDIAN_INFO,
+            InitLogMessage::PIEnclaveFullyInitialized { .. } => Self::PI_FULLY_INITIALIZED,
+            InitLogMessage::OAActivated { .. } => Self::OA_ACTIVATED,
+        };
+
+        Self::object_key_for_suffix(session_id, suffix)
+    }
+
+    pub fn object_key_pattern(&self, session_id: &str) -> ObjectKeyPattern {
+        ObjectKeyPattern::Fixed(self.object_key(session_id))
+    }
+
+    pub fn attestation_object_key(session_id: &str) -> String {
+        Self::object_key_for_suffix(session_id, Self::OI_ATTEST_UNSIGNED)
+    }
+
+    pub fn guardian_info_object_key(session_id: &str) -> String {
+        Self::object_key_for_suffix(session_id, Self::OI_GUARDIAN_INFO)
+    }
+
+    fn object_key_for_suffix(session_id: &str, suffix: &str) -> String {
+        format!("{S3_DIR_INIT}/{session_id}/{suffix}.json")
+    }
+}
